requirements-audio.txt
openai-whisper
pyttsx3
sounddevice
SpeechRecognition
